feat: create an ondemand render for settings
