+++ src/domains/settings/pages/Settings/Settings.stories.tsx
@@ -6,19 +6,19 @@ export default { title: "Pages / Settings", decorators: [withKnobs] };
 
 const items = [
 	{
-		itemKey: "general",
+		itemKey: "General",
 		label: "General",
 		icon: "general",
 		route: "/settings/general",
 	},
 	{
-		itemKey: "peer",
+		itemKey: "Peer",
 		label: "Peer",
 		icon: "peer",
 		route: "/settings/peer",
 	},
 	{
-		itemKey: "plugins",
+		itemKey: "Plugins",
 		label: "Plugins",
 		icon: "plugin",
 		route: "/settings/plugins",
@@ -33,7 +33,7 @@ const settingsProps = {
 };
 
 export const Default = () => {
-	const [activeSettings, setActiveSettings] = useState("general");
+	const [activeSettings, setActiveSettings] = useState("General");
 
 	return (
 		<div className="w-full h-full">
