+++ src/domains/settings/pages/Settings/available-settings/index.ts
@@ -1,0 +1,1 @@
+export * from "./General.tsx";
